requirements.txt
asgiref==3.9.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.1
Django==5.2.6
djangorestframework==3.16.1
dotenv==0.9.9
idna==3.10
josepy==2.1.0
mozilla-django-oidc==4.0.1
pycparser==2.23
python-dotenv==1.1.1
requests==2.32.5
sqlparse==0.5.3
urllib3==2.5.0
